fix mac os warn

+++ src/Common/Network/NetConnect.cpp
@@ -9,12 +9,12 @@ NetConnect::~NetConnect()
 {
 }
 
-int32 NetConnect::Send(const void * data, int32 len)
+int64 NetConnect::Send(const void * data, int32 len)
 {
 	return send(m_socket, (char*)data, len, 0);
 }
 
-int32 NetConnect::Recv(void * data, int32 len)
+int64 NetConnect::Recv(void * data, int32 len)
 {
 	return recv(m_socket, (char*)data, len, 0);
 }
