+++ src/Common/Network/NetConnect.h
@@ -6,8 +6,8 @@ public:
 	NetConnect();
 	virtual ~NetConnect();
 
-	virtual int32 Send(const void * data, int32 len);
-	virtual int32 Recv(void * data, int32 len);
+	virtual int64 Send(const void * data, int32 len);
+	virtual int64 Recv(void * data, int32 len);
 };
 
 YGAME_SERVER_END
